Tests: moved created tests to the existing Document test framework.
